Refactor render payloads around sim snapshots
Switch the render handoff from render_frame_payload to sim_snapshot,
track fixed-step time in nanoseconds, and split frame UBO data into
separate sim and render timing fields.

Also move frame UBO uploads to end_render_frame and fix Windows
nanosecond conversion in the profile timer.

+++ src/profile_timer.h
@@ -50,7 +50,7 @@ struct profile_timepoints
     s64 dt_ns;
 
     // Time, in s, between current split point and previous split (updated with ptimer_split)
-    f64 dt;
+    f32 dt;
 };
 
 inline f64 nanos_to_sec(s64 ns) {
